pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "network-test-automation-framework"
version = "1.0.0"
description = "Production-grade, vendor-agnostic network test automation framework with multi-vendor support"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Network Automation Team" },
]
keywords = [
    "network",
    "automation",
    "testing",
    "juniper",
    "cisco",
    "arista",
    "evpn",
    "bgp",
    "batfish",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
juniper = [
    "junos-eznc>=2.7",
    "jsnapy>=1.3",
]
cisco = [
    "napalm>=4.0",
    "netmiko>=4.0",
]
arista = [
    "pyeapi>=1.0",
    "napalm>=4.0",
]
all-vendors = [
    "junos-eznc>=2.7",
    "jsnapy>=1.3",
    "napalm>=4.0",
    "netmiko>=4.0",
    "pyeapi>=1.0",
]
nornir = [
    "nornir>=3.4",
    "nornir-utils>=0.2",
    "nornir-napalm>=0.4",
]
batfish = [
    "pybatfish>=2023.10.0",
]
traffic = [
    "ixnetwork-restpy>=1.1",
]
triage = [
    "anthropic>=0.25",
]
reporting = [
    "jinja2>=3.1",
    "weasyprint>=60.0",
]
robot = [
    "robotframework>=7.0",
]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "black>=24.0",
    "pre-commit>=3.7",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.5",
    "pytest-mock>=3.14",
    "pytest-timeout>=2.3",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[project.urls]
Homepage = "https://github.com/example/network-test-automation-framework"
Repository = "https://github.com/example/network-test-automation-framework"
Documentation = "https://example.github.io/network-test-automation-framework"
Issues = "https://github.com/example/network-test-automation-framework/issues"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "-m", "not integration",
]
markers = [
    "integration: marks tests requiring live device connectivity",
    "slow: marks tests that take a long time to run",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "D",     # pydocstyle
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
]
ignore = [
    "D100",  # Missing docstring in public module (covered by __init__.py)
    "D104",  # Missing docstring in public package
    "D203",  # One blank line before class (conflicts with D211)
    "D213",  # Multi-line summary second line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S101"]
"**/conftest.py" = ["D"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "jnpr.*",
    "napalm.*",
    "netmiko.*",
    "pyeapi.*",
    "nornir.*",
    "pybatfish.*",
    "ixnetwork_restpy.*",
    "jsnapy.*",
    "anthropic.*",
    "weasyprint.*",
    "yaml.*",
]
ignore_missing_imports = true

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]
